+++ app/views/conversations/_assistant_sidebar_link.html.erb
@@ -12,13 +12,13 @@
                 variant: :micro,
                 size: 18,
                 class: "outline-none inline text-gray-400 invisible group-hover:visible hover:text-white cursor-pointer",
-                'data-controller': "nested-clickable",
+                'data-controller': "nested-pointer",
                 title: "More",
                 tabindex: 0,
                 role: :button
           %>
 
-          <ul tabindex="0" class="dropdown-content z-10 menu p-2 shadow-xl bg-base-100 rounded-box w-52 -mr-10 mt-7 dark:!bg-zinc-800" data-controller="nested-clickable">
+          <ul tabindex="0" class="dropdown-content z-10 menu p-2 shadow-xl bg-base-100 rounded-box w-52 -mr-10 mt-7 dark:!bg-zinc-800" data-controller="nested-pointer">
             <li><a><%= icon "cog-6-tooth", variant: :outline, size: 18 %>Settings</a></li>
             <li><a><%= icon "arrow-up-tray", variant: :outline, size: 18 %>Share</a></li>
           </ul>
@@ -29,7 +29,7 @@
         variant: :micro,
         size: 18,
         class: "inline text-gray-400 cursor-pointer #{defined?(selected) && selected ? 'text-white' : 'invisible group-hover:visible hover:text-white'}",
-        'data-controller': "nested-clickable",
+        'data-controller': "nested-pointer",
         title: "New chat"
       %>
     </div>
